pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clasi"
version = "0.20260330.2"
description = "MCP server for AI-driven software engineering process — agents, skills, instructions, and artifact management"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Eric Busboom"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "claude-agent-sdk>=0.1",
    "click>=8.0",
    "jinja2>=3.0",
    "jsonschema>=4.0",
    "mcp>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
clasi = "clasi.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/e2e/project", "tests/e2e/repro_project"]
addopts = "--cov=clasi --cov-report=term-missing --cov-report=lcov:lcov.info -m 'not slow'"
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.setuptools.package-data]
clasi = [
    "agents/**/*.md",
    "agents/**/*.md.j2",
    "agents/**/*.yaml",
    "contract-schema.yaml",
    "init/*.md",
    "skills/*.md",
    "instructions/*.md",
    "instructions/languages/*.md",
    "rules/*.md",
    "templates/*.md",
    "se-overview-template.md",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["clasi*"]
